+++ R/namedMapBuilder.R
@@ -105,6 +105,9 @@ early_tries <- function(nm, vl, values) {
 #' @export
 `:=.factor` <- named_map_builder
 
+#' @export
+`:=.logical` <- named_map_builder
+
 
 
 
@@ -135,6 +138,8 @@ early_tries <- function(nm, vl, values) {
 #' @export
 `%:=%.factor` <- named_map_builder
 
+#' @export
+`%:=%.logical` <- named_map_builder